requirements.txt
python-telegram-bot
#sqlite3
pysqlite3
#para levantar con doker